% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_exists}
\alias{cache_exists}
\title{Check if Data is Cached}
\usage{
cache_exists(filename)
}
\arguments{
\item{filename}{Name of the cached file.}
}
\value{
Logical indicating if file exists in cache.
}
\description{
Internal function to check if data exists in cache.
}
\keyword{internal}
