% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_save}
\alias{cache_save}
\title{Save Data to Cache}
\usage{
cache_save(data, filename)
}
\arguments{
\item{data}{Data to cache.}

\item{filename}{Name for the cached file.}
}
\value{
Invisible path to the cached file.
}
\description{
Internal function to save data to the local cache.
}
\keyword{internal}
