% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_leaderboard.R
\name{fetch_leaderboard_fast}
\alias{fetch_leaderboard_fast}
\title{Fast leaderboard fetch using site API}
\usage{
fetch_leaderboard_fast(event_id, year, tour = "pga")
}
\description{
Fast leaderboard fetch using site API
}
\keyword{internal}
