% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_holes.R
\name{fetch_player_holes}
\alias{fetch_player_holes}
\title{Fetch hole-by-hole for a single player}
\usage{
fetch_player_holes(event_id, player_id, tour = "pga")
}
\description{
Fetch hole-by-hole for a single player
}
\keyword{internal}
