% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_espn.R
\name{get_espn_tour_code}
\alias{get_espn_tour_code}
\title{Get ESPN tour code}
\usage{
get_espn_tour_code(tour = "pga")
}
\arguments{
\item{tour}{Tour identifier ("pga", "lpga", "euro", etc.)}
}
\value{
ESPN tour code string
}
\description{
Get ESPN tour code
}
\keyword{internal}
