% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_majors}
\alias{get_majors}
\title{Get Major Championships}
\usage{
get_majors(year, db_path = "data/golfastr.duckdb")
}
\arguments{
\item{year}{Season year}

\item{db_path}{Path to database file}
}
\value{
Tibble with major championship results
}
\description{
Get results from the four major championships.
}
\examples{
\dontrun{
# Get 2025 majors
get_majors(2025)
}
}
