% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dictionary.R
\name{pga_majors}
\alias{pga_majors}
\title{Get PGA Major Championships}
\usage{
pga_majors()
}
\value{
A tibble with tournament, month, and course columns.
}
\description{
Returns a data frame with the four major championships
and their typical schedule.
}
\examples{
pga_majors()
}
