% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_leaderboard}
\alias{plot_leaderboard}
\title{Plot Tournament Leaderboard}
\usage{
plot_leaderboard(
  tournament,
  year,
  top_n = 10,
  db_path = "data/golfastr.duckdb"
)
}
\arguments{
\item{tournament}{Tournament name}

\item{year}{Season year}

\item{top_n}{Number of players to show}

\item{db_path}{Path to database}
}
\value{
ggplot object
}
\description{
Bar chart of tournament leaderboard.
}
