% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_scoring}
\alias{plot_scoring}
\title{Plot Scoring Distribution}
\usage{
plot_scoring(tournament, year, db_path = "data/golfastr.duckdb")
}
\arguments{
\item{tournament}{Tournament name}

\item{year}{Season year}

\item{db_path}{Path to database}
}
\value{
ggplot object
}
\description{
Histogram of tournament scores.
}
