% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_wins}
\alias{plot_wins}
\title{Plot Win Distribution}
\usage{
plot_wins(year = NULL, top_n = 10, db_path = "data/golfastr.duckdb")
}
\arguments{
\item{year}{Optional year filter}

\item{top_n}{Number of players to show}

\item{db_path}{Path to database}
}
\value{
ggplot object
}
\description{
Pie/bar chart of wins by player.
}
