% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_winners}
\alias{get_winners}
\title{Get Tournament Winners}
\usage{
get_winners(year = NULL, db_path = "data/golfastr.duckdb")
}
\arguments{
\item{year}{Optional year filter}

\item{db_path}{Path to database file}
}
\value{
Tibble with tournament winners
}
\description{
Get all tournament winners from the database.
}
\examples{
\dontrun{
# Get all winners
get_winners()

# Get 2025 winners
get_winners(2025)
}
}
