% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golfastr-package.R
\docType{package}
\name{golfastr-package}
\alias{golfastr}
\alias{golfastr-package}
\title{golfastr: Efficiently Access Pro Golf Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Fetch PGA Tour tournament data from ESPN including leaderboards and hole-by-hole scoring. Data is returned in tidy tibble format ready for analysis. Supports local DuckDB storage for fast repeated access.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/array-carpenter/golfastr}
  \item Report bugs at \url{https://github.com/array-carpenter/golfastr/issues}
}

}
\author{
\strong{Maintainer}: Ray Carpenter \email{raymondcarpenter1@gmail.com}

}
\keyword{internal}
