% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_storage.R
\name{load_from_rds}
\alias{load_from_rds}
\title{Load Tournament Data from RDS}
\usage{
load_from_rds(file_path = "data/golfastr.rds")
}
\arguments{
\item{file_path}{Path to RDS file}
}
\value{
Tibble with tournament data
}
\description{
Load Tournament Data from RDS
}
