% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{made_cuts_leaders}
\alias{made_cuts_leaders}
\title{Made Cuts Percentage}
\usage{
made_cuts_leaders(
  year = NULL,
  min_events = 5,
  top_n = 10,
  db_path = "data/golfastr.duckdb"
)
}
\arguments{
\item{year}{Optional year filter}

\item{min_events}{Minimum events played}

\item{top_n}{Number of players to show}

\item{db_path}{Path to database}
}
\value{
Tibble with made cuts leaders
}
\description{
Get players by percentage of cuts made.
}
