% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_leaderboard}
\alias{plot_leaderboard}
\title{Plot Tournament Leaderboard}
\usage{
plot_leaderboard(tournament, year, top_n = 10, file_path)
}
\arguments{
\item{tournament}{Tournament name}

\item{year}{Season year}

\item{top_n}{Number of players to show}

\item{file_path}{Path to data file (.rds or .parquet)}
}
\value{
ggplot object
}
\description{
Bar chart of tournament leaderboard.
}
\examples{
\dontrun{
plot_leaderboard("Masters", 2025, file_path = "golf_data.rds")
}
}
