% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_scoring}
\alias{plot_scoring}
\title{Plot Scoring Distribution}
\usage{
plot_scoring(tournament, year, file_path)
}
\arguments{
\item{tournament}{Tournament name}

\item{year}{Season year}

\item{file_path}{Path to data file (.rds or .parquet)}
}
\value{
ggplot object
}
\description{
Histogram of tournament scores.
}
\examples{
\dontrun{
plot_scoring("Masters", 2025, file_path = "golf_data.rds")
}
}
