% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_wins}
\alias{plot_wins}
\title{Plot Win Distribution}
\usage{
plot_wins(year = NULL, top_n = 10, file_path)
}
\arguments{
\item{year}{Optional year filter}

\item{top_n}{Number of players to show}

\item{file_path}{Path to data file (.rds or .parquet)}
}
\value{
ggplot object
}
\description{
Pie/bar chart of wins by player.
}
\examples{
\dontrun{
plot_wins(year = 2025, file_path = "golf_data.rds")
}
}
