% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_storage.R
\name{save_to_rds}
\alias{save_to_rds}
\title{Save Tournament Data to RDS}
\usage{
save_to_rds(data, file_path = "data/golfastr.rds", append = TRUE)
}
\arguments{
\item{data}{Tibble of tournament data}

\item{file_path}{Path to RDS file (default: "data/golfastr.rds")}

\item{append}{If TRUE, append to existing data}
}
\description{
Save tournament data to an RDS file.
}
